refactor: Adjust namespaces in the Tests project.
