+++ src/Syringe.Tests/Unit/Web/Controllers/TestControllerTests.cs
@@ -9,7 +9,7 @@ using Syringe.Web.Controllers;
 using Syringe.Web.Mappers;
 using Syringe.Web.Models;
 
-namespace Syringe.Tests.Unit.Web
+namespace Syringe.Tests.Unit.Web.Controllers
 {
     [TestFixture]
     public class TestControllerTests
