+++ src/Syringe.Tests/Unit/Web/ValidationAttributes/ValidationAttributesTests.cs
@@ -1,8 +1,7 @@
-﻿using System.ComponentModel.DataAnnotations;
-using NUnit.Framework;
+﻿using NUnit.Framework;
 using Syringe.Web.ValidationAttributes;
 
-namespace Syringe.Tests.Unit.Web
+namespace Syringe.Tests.Unit.Web.ValidationAttributes
 {
     [TestFixture]
     public class ValidRegexAttributeTests
